Windows build warning fix
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7510 3db68df2-c116-0410-a063-a993310a9797

+++ source/scriptinterface/ScriptInterface.cpp
@@ -548,7 +548,7 @@ void* ScriptInterface::GetPrivate(JSContext* cx, JSObject* obj)
 void ScriptInterface::DumpHeap()
 {
 #ifdef DEBUG
-	JS_DumpHeap(m->m_cx, stderr, NULL, 0, NULL, -1, NULL);
+	JS_DumpHeap(m->m_cx, stderr, NULL, 0, NULL, (size_t)-1, NULL);
 #endif
 	fprintf(stderr, "# Bytes allocated: %d\n", JS_GetGCParameter(m->m_rt, JSGC_BYTES));
 	JS_GC(m->m_cx);
